backend/requirements.txt
# EBA XBRL Validator Backend Dependencies
# Python 3.11 compatible packages with pinned versions

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0

# YAML configuration support
PyYAML==6.0.1

# Process and system utilities
psutil==5.9.6

# HTTP client (for future use, currently blocked in offline mode)
httpx==0.25.2

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Arelle XBRL processor
# Integrated as git submodule under third_party/arelle/
# Version: edgr19.2.1 (commit: 5406dd6bf202f705b2aead68b7215ebf0db3b317)

# Additional utilities
pathlib2==2.3.7; python_version < "3.4"

# Required by Arelle
regex
lxml
isodate
Pillow

# Arelle core runtime dependencies (aligned with vendored Arelle requirements)
bottle==0.13.4
cheroot==11.0.0
filelock==3.19.1
jsonschema==4.25.1
numpy==2.0.2; python_version <= "3.9"
openpyxl==3.1.5
pycountry==24.6.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
tornado==6.5.2
typing-extensions==4.15.0
tinycss2==1.4.0
